requirements.txt
streamlit==1.37.0
tqdm
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.9.1
sentence-transformers==3.0.1
faiss-cpu==1.8.0
PyPDF2==3.0.1
python-pptx==1.0.2
openai>=1.0.0
python-dotenv==1.0.1
torch==2.2.2
scipy==1.11.4
joblib==1.3.2